% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable-nodes.R
\name{bsm_names_variable_nodes}
\alias{bsm_names_variable_nodes}
\title{Names Variable Nodes}
\usage{
bsm_names_variable_nodes(x, stochastic = NA)
}
\arguments{
\item{x}{A string of JAGS model code}

\item{stochastic}{A flag specifying whether to get stochastic (TRUE), 
deterministic (FALSE) or both types (NA) of variable node.}
}
\value{
A sorted character vector of the variable node names.
}
\description{
Gets names of variable (as opposed to constant) nodes in JAGS model code.
}
