% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{bsm_simulate_data}
\alias{bsm_simulate_data}
\title{Simulate Data}
\usage{
bsm_simulate_data(jags_code, constants, parameters, monitor = ".*")
}
\arguments{
\item{jags_code}{A string of the jags code to simulate the data.
It must not be in a data or model block.}

\item{constants}{A uniquely named list of numeric vectors, matrices and arrays
specifying constant values.}

\item{parameters}{A uniquely named list of numeric vectors, matrices and arrays
specifying parameter values.}

\item{monitor}{A regular expression specifying the data to}
}
\value{
A uniquely named list of numeric vectors, matrices and arrays of the 
simulated data values.
}
\description{
The jags codes should not be in a block.
The constant and parameter arguments cannot include missing values.
}
