% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-comments.R
\name{bsm_strip_comments}
\alias{bsm_strip_comments}
\title{Strip Comments}
\usage{
bsm_strip_comments(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
The modified character vector.
}
\description{
Strips all \code{#} comments from a character vector.
}
\examples{
bsm_strip_comments("x <- 1 # a comment")
}
