% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place-holder.R
\name{place_holder}
\alias{place_holder}
\title{Place Holder}
\usage{
place_holder()
}
\value{
A flag.
}
\description{
A temporary function to act as a place holder for setting up the structure.
}
\examples{
place_holder()
}
