% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{sims_add}
\alias{sims_add}
\title{Add Simulated Datasets}
\usage{
sims_add(path = ".", nsims = 1)
}
\arguments{
\item{path}{A string specifying the path to the directory to
add the data sets to.}

\item{nsims}{A count of the number of additional datasets to generate.}
}
\value{
A character vector of the names of the files created.
}
\description{
Add Simulated Datasets
}
