% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-distributions.R
\name{sims_add_distributions}
\alias{sims_add_distributions}
\title{Sims Add Distributions}
\usage{
sims_add_distributions(x = character(0))
}
\arguments{
\item{x}{A character vector of the random variate generating R functions to add.}
}
\value{
A character vector of the names of .
}
\description{
Sims Add Distributions
}
\examples{
print(sims_add_distributions("llog"))
}
