% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-args.R
\name{sims_args}
\alias{sims_args}
\title{Simulated Data Arguments}
\usage{
sims_args(path = "sims")
}
\arguments{
\item{path}{A string of the path to the directory with the simulated data.}
}
\value{
A list of the values in \code{file.path(path, '.sims_args.rds')}.
}
\description{
Checks that the number and names of the data files are consistent with the 
number of simulations in the '.sims_args.rds' file.
}
\examples{
set.seed(10)
sims_simulate("a ~ dunif(0,1)", nsims = 2L, 
  write = TRUE, path = tempdir(), exists = NA)
sims_args(tempdir())
}
