% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{sims_check}
\alias{sims_check}
\title{Check Simulated Data}
\usage{
sims_check(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the simulated data.}
}
\value{
An informative error or invisible list of the argument values.
}
\description{
Checks the simulated data argument values in the '.sims.rds' file.
}
\details{
The checks include whether number and names of the data files
in the directory are consistent with the number of simulations.
}
\examples{
set.seed(10)
sims_simulate("a <- runif(1)",
  save = TRUE, path = tempdir(), exists = NA,
  ask = FALSE
)
(sims_check(tempdir()))
}
