% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-distributions.R
\name{sims_distributions}
\alias{sims_distributions}
\title{Sims Distributions}
\usage{
sims_distributions()
}
\value{
A character vector of the names of random variate generating R functions.
}
\description{
Gets names of the random variate generating R functions recognized by sims 
as producing stochastic variable nodes.
}
\details{
By default the functions are the random variate generating functions
listed in \code{\link{Distributions}}.
}
\examples{
sims_distributions()
}
