% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sims_rcount}
\alias{sims_rcount}
\title{Random Count}
\usage{
sims_rcount(n = 1L)
}
\arguments{
\item{n}{A count of the number of random counts to generate.}
}
\value{
A random count.
}
\description{
Generates a platform independent random count between 0 and 
2,147,483,647.
It is useful for specifying values for \code{\link{set.seed}()}.
}
\examples{
set.seed(45)
sims_rcount()
set.seed(45)
sims_rcount(n = 2L)
}
