% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_rdists}
\alias{sims_rdists}
\title{Sims Random R Distributions}
\usage{
sims_rdists()
}
\value{
A character vector.
}
\description{
Gets the names of the R random variate generating functions
listed in \code{\link[=Distributions]{Distributions()}}.
}
\examples{
sims_rdists()
}
