% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_rdists_reset}
\alias{sims_rdists_reset}
\title{Sims Reset R Distributions}
\usage{
sims_rdists_reset()
}
\value{
A invisible character vector of the names of the previous R functions.
}
\description{
Resets the names of the R functions recognized by sims
as producing stochastic variable nodes to those listed in \code{\link{Distributions}}.
}
\examples{
print(sims_rdists_reset())
}
\seealso{
\code{\link{sims_rdists}()} and
\code{\link{sims_rdists_set}()}
}
