% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_rdists_set}
\alias{sims_rdists_set}
\title{Sims Set Distributions}
\usage{
sims_rdists_set(x)
}
\arguments{
\item{x}{A character vector of the names of the R functions.}
}
\value{
A invisible character vector of the names of the previous R functions.
}
\description{
Sets the names of the R functions recognized by sims
as producing stochastic variable nodes.
}
\examples{
print(sims_rdists_set("llog"))
}
\seealso{
\code{\link{sims_rdists}()} and
\code{\link{sims_rdists_reset}()}
}
