% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-distributions.R
\name{sims_reset_distributions}
\alias{sims_reset_distributions}
\title{Sims Reset Distributions}
\usage{
sims_reset_distributions()
}
\value{
A invisible character vector of the names of the previous R distributions.
}
\description{
Resets sims distributions.
}
\examples{
print(sims_reset_distributions())
}
\seealso{
\code{\link{sims_distributions}()} and  
\code{\link{sims_set_distributions}()}
}
