% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_reset_rdists}
\alias{sims_reset_rdists}
\title{Sims Reset Distributions}
\usage{
sims_reset_rdists()
}
\value{
A invisible character vector of the names of the previous R distributions.
}
\description{
Resets sims distributions.
}
\examples{
print(sims_reset_rdists())
}
\seealso{
\code{\link{sims_rdists}()} and  
\code{\link{sims_set_rdists}()}
}
