% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-distributions.R
\name{sims_set_distributions}
\alias{sims_set_distributions}
\title{Sims Set Distributions}
\usage{
sims_set_distributions(x)
}
\arguments{
\item{x}{A character vector of the random variate generating R functions.}
}
\value{
A invisible character vector of the names of the previous R distributions.
}
\description{
Sets sims distributions.
}
\examples{
print(sims_set_distributions("llog"))
}
\seealso{
\code{\link{sims_distributions}()} and  
\code{\link{sims_reset_distributions}()}
}
