% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_set_rdists}
\alias{sims_set_rdists}
\title{Sims Set Distributions}
\usage{
sims_set_rdists(x)
}
\arguments{
\item{x}{A character vector of the random variate generating R functions.}
}
\value{
A invisible character vector of the names of the previous R distributions.
}
\description{
Sets sims distributions.
}
\examples{
print(sims_set_rdists("llog"))
}
\seealso{
\code{\link{sims_rdists}()} and  
\code{\link{sims_reset_rdists}()}
}
