% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{bsm_simulate_data}
\alias{bsm_simulate_data}
\title{Simulate Data}
\usage{
bsm_simulate_data(code, fixed = named_list(), monitor = ".*",
  nsamples = 100L, silent = FALSE)
}
\arguments{
\item{code}{A string of the JAGS code to simulate the data.
The code must not be in a data or model block.}

\item{fixed}{A uniquely named list of numeric vectors, matrices and arrays
specifying known non-missing values. 
The names cannot match variable (stochastic or deterministic) nodes in code.}

\item{monitor}{A character vector specifying the names of the stochastic nodes 
to simulate or if a string (length one character vector) a regular expression 
that must match at least one stochastic node in code.}

\item{nsamples}{A positive integer of the number of samples to simulate.}

\item{silent}{A flag specifying whether to suppress messages and warnings.}
}
\value{
The simulated data values in the form of a single chain 
\code{\link[mcmcr]{mcmcr}} object where each iteration represents one sample.
}
\description{
Simulate Data
}
\examples{
set.seed(101)
bsm_simulate_data("a ~ dunif(0, 1)", nsamples = 1L)
}
