% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{sims_add}
\alias{sims_add}
\title{Add Simulated Datasets}
\usage{
sims_add(path = ".", nsims = 1, progress = FALSE,
  options = furrr::future_options())
}
\arguments{
\item{path}{A string specifying the path to the directory to add the data sets to.}

\item{nsims}{A count of the number of additional datasets to generate.
By default 100 datasets are added.}

\item{progress}{A flag specifying whether to print a progress bar.}

\item{options}{The future specific options to use with the workers.}
}
\value{
A character vector of the names of the files created
(and in the case of argsism, created).
}
\description{
Add Simulated Datasets
}
