% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{sims_add}
\alias{sims_add}
\title{Add Simulated Datasets}
\usage{
sims_add(
  path = ".",
  nsims = 1,
  progress = FALSE,
  options = furrr::future_options()
)
}
\arguments{
\item{path}{A string specifying the path to the directory to add the data sets to.}

\item{nsims}{A count of the number of additional datasets to generate.}

\item{progress}{A flag specifying whether to print a progress bar.}

\item{options}{The furrr specific options to use with the workers.
The seed should be specified using \code{\link[base:set.seed]{base::set.seed()}}.}
}
\value{
A character vector of the names of the files created
(and in the case of argsism, created).
}
\description{
Add Simulated Datasets
}
