% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-add.R
\name{sims_add}
\alias{sims_add}
\title{Add Simulated Datasets}
\usage{
sims_add(nsims = getOption("sims.nsims", 100L), path = "sims")
}
\arguments{
\item{nsims}{A count of the number of additional datasets to generate.
By default 100 datasets are added.}

\item{path}{A string specifying the path to the directory to add the data sets to.}
}
\value{
A character vector of the names of the files created 
(and in the case of argsism, created).
}
\description{
Add Simulated Datasets
}
