% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-args.R
\name{sims_args}
\alias{sims_args}
\title{Simulated Data Argument Values}
\usage{
sims_args(path)
}
\arguments{
\item{path}{A string of the path to the directory with the simulated data.}
}
\value{
A named list of the values in \code{file.path(path, '.sims_args.rds')}.
}
\description{
Gets the simulated data argument values in the '.sims_args.rds' file.
}
\examples{
set.seed(10)
sims_simulate("a ~ dunif(0,1)", path = tempdir(), exists = NA)
sims_args(tempdir())
}
