% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-check.R
\name{sims_check}
\alias{sims_check}
\title{Check Simulated Data}
\usage{
sims_check(path = "sims")
}
\arguments{
\item{path}{A string of the path to the directory with the simulated data.}
}
\value{
A list of the values in \code{file.path(path, '.argsims.rds')}.
}
\description{
Checks that the number and names of the data files are consistent with the 
number of simulations in the '.argsims.rds' file.
}
