% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-check.R
\name{sims_check}
\alias{sims_check}
\title{Check Simulated Data}
\usage{
sims_check(dir = "sims", path = ".")
}
\arguments{
\item{dir}{A string of the name of the directory with the simulated data.}

\item{path}{A string of the path to \code{dir}.}
}
\value{
A invisible copy of a list of the values in 'argsims.rds'
if all checks are passed. Otherwise it throws an error.
}
\description{
Checks that \code{dir} includes an 'argsims.rds' file and 
that the number and names of the data files are consistent with the 
number of simulations.
}
