% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-copy.R
\name{sims_copy}
\alias{sims_copy}
\title{Copy Simulated Datasets}
\usage{
sims_copy(dir_from = "sims", path_from = ".",
  dir_to = paste0(dir_from, "_copy"), path_to = path_from,
  exists = FALSE)
}
\arguments{
\item{dir_from}{A string of the name of the directory containing the simulated datasets.}

\item{path_from}{A string of the path to \code{dir_from}.}

\item{dir_to}{A string of the name of the directory to copy the simulated dataset to.}

\item{path_to}{A string of the path to \code{dir_to}.}

\item{exists}{A flag specifying whether \code{dir_to} should already exist.
If \code{exists = NA} it doesn't matter. If the directory already exists it is 
overwritten if \code{exists = TRUE} or \code{exists = NA} otherwise an
error is thrown.}
}
\value{
A string of the path to the new directory, ie \code{file.path(path_to, dir_to)}.
}
\description{
Copy Simulated Datasets
}
