% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims-copy.R
\name{sims_copy}
\alias{sims_copy}
\title{Copy Simulated Datasets}
\usage{
sims_copy(path_from = "sims", path_to = paste0(path_from, "_copy"),
  exists = FALSE)
}
\arguments{
\item{path_from}{A string of the path to the directory containing the simulated datasets.}

\item{path_to}{A string of the path to the directory to copy the simulated dataset to.}

\item{exists}{A flag specifying whether \code{path_to} should already exist.
If \code{exists = NA} it doesn't matter. If the directory already exists it is 
overwritten if \code{exists = TRUE} or \code{exists = NA} otherwise an
error is thrown.}
}
\value{
A character vector of the names of the files copied.
}
\description{
Does not check the sims.
}
