% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{sims_copy}
\alias{sims_copy}
\title{Copy Simulated Datasets}
\usage{
sims_copy(
  path_from = ".",
  path_to = paste0(path_from, "_copy"),
  exists = FALSE,
  ask = getOption("sims.ask", TRUE),
  silent = FALSE
)
}
\arguments{
\item{path_from}{A string of the path to the directory containing
the simulated datasets.}

\item{path_to}{A string of the path to the directory to copy
the simulated dataset to.}

\item{exists}{A flag specifying whether \code{path_to} should already exist.
If \code{exists = NA} it doesn't matter. If the directory already exists
sims compatible files are deleted if \code{exists = TRUE} or \code{exists = NA}
otherwise an
error is thrown.}

\item{ask}{A flag specifying whether to ask before deleting files.}

\item{silent}{A flag specifying whether to suppress warnings.}
}
\value{
A character vector of the names of the files copied.
}
\description{
Copy Simulated Datasets
}
