% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-files.R
\name{sims_data_files}
\alias{sims_data_files}
\title{Simulated Data Files}
\usage{
sims_data_files(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the simulated data.}
}
\value{
A character vector of the names of the simulated data files.
}
\description{
Gets the names of the simulated data files.
}
\examples{
set.seed(10)
sims_simulate("a <- runif(1)",
  nsims = 10L, path = tempdir(),
  exists = NA, ask = FALSE
)
sims_data_files(tempdir())
}
