% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{sims_info}
\alias{sims_info}
\title{Simulated Data Argument Values}
\usage{
sims_info(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the simulated data.}
}
\value{
A named list of the values in \code{file.path(path, '.sims.rds')}.
}
\description{
Gets the simulated data argument values in the '.sims.rds' file.
}
\examples{
set.seed(10)
sims_simulate("a <- runif(1)", path = tempdir(), exists = NA, ask = FALSE)
sims_info(tempdir())
}
