% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_rdists}
\alias{sims_rdists}
\title{Sims Random R Distributions}
\usage{
sims_rdists()
}
\value{
A character vector of the names of random variate generating R functions.
}
\description{
Gets the names of the random variate generating R functions recognized by sims 
as producing stochastic variable nodes.
}
\details{
By default the functions are the random variate generating functions
listed in \code{\link{Distributions}}.
}
\examples{
sims_rdists()
}
\seealso{
\code{\link{sims_add_rdists}()} and  
\code{\link{sims_reset_rdists}()}
}
