% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdists.R
\name{sims_rdists_add}
\alias{sims_rdists_add}
\title{Sims Add Distributions}
\usage{
sims_rdists_add(x)
}
\arguments{
\item{x}{A character vector of the random variate generating R functions.}
}
\value{
A invisible character vector of the names of the previous R distributions.
}
\description{
Adds sims distributions.
}
\examples{
print(sims_rdists_add("llog"))
}
\seealso{
\code{\link{sims_rdists}()} and  
\code{\link{sims_rdists_reset}()}
}
